MBX-2530 Moved Settings model
