+++ Mindbox/InAppMessages/Models/Config/SettingsModel.swift
@@ -1,0 +1,24 @@
+//
+//  SettingsModel.swift
+//  Mindbox
+//
+//  Created by vailence on 15.06.2023.
+//  Copyright © 2023 Mindbox. All rights reserved.
+//
+
+import Foundation
+
+struct Settings: Decodable, Equatable {
+    let operations: SettingsOperations?
+    
+    struct SettingsOperations: Decodable, Equatable {
+        
+        let viewProduct: Operation?
+        let viewCategory: Operation?
+        let setCart: Operation?
+        
+        struct Operation: Decodable, Equatable {
+            let systemName: String
+        }
+    }
+}
