+++ MindboxTests/InAppMessages/Tests/InAppConfigResponseTests/InAppConfigStub.swift
@@ -13,7 +13,7 @@ final class InAppConfigStub {
     func getCategoryID_Substring() -> InAppConfigResponse {
         let categoryIdTargeting = CategoryIDTargeting(kind: .substring, value: "oot")
         let targeting = Targeting.viewProductCategoryId(categoryIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -22,7 +22,7 @@ final class InAppConfigStub {
     func getCategoryID_notSubstring() -> InAppConfigResponse {
         let categoryIdTargeting = CategoryIDTargeting(kind: .notSubstring, value: "oot")
         let targeting = Targeting.viewProductCategoryId(categoryIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -31,7 +31,7 @@ final class InAppConfigStub {
     func getCategoryID_startWith() -> InAppConfigResponse {
         let categoryIdTargeting = CategoryIDTargeting(kind: .startsWith, value: "oot")
         let targeting = Targeting.viewProductCategoryId(categoryIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -40,7 +40,7 @@ final class InAppConfigStub {
     func getCategoryID_endWith() -> InAppConfigResponse {
         let categoryIdTargeting = CategoryIDTargeting(kind: .endsWith, value: "ots")
         let targeting = Targeting.viewProductCategoryId(categoryIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -51,7 +51,7 @@ final class InAppConfigStub {
                                                           values: [.init(id: "testik2", name: "System1C"),
                                                                    .init(id: "81", name: "TestSite")])
         let targeting = Targeting.viewProductCategoryIdIn(categoryIdInTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -62,7 +62,7 @@ final class InAppConfigStub {
                                                           values: [.init(id: "testik2", name: "System1C"),
                                                                    .init(id: "81", name: "TestSite")])
         let targeting = Targeting.viewProductCategoryIdIn(categoryIdInTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -81,7 +81,7 @@ final class InAppConfigStub {
     func getProductID_Substring() -> InAppConfigResponse {
         let productIdTargeting = ProductIDTargeting(kind: .substring, value: "oot")
         let targeting = Targeting.viewProductId(productIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -90,7 +90,7 @@ final class InAppConfigStub {
     func getProductID_notSubstring() -> InAppConfigResponse {
         let productIdTargeting = ProductIDTargeting(kind: .notSubstring, value: "oot")
         let targeting = Targeting.viewProductId(productIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -99,7 +99,7 @@ final class InAppConfigStub {
     func getProductID_startsWith() -> InAppConfigResponse {
         let productIdTargeting = ProductIDTargeting(kind: .startsWith, value: "oot")
         let targeting = Targeting.viewProductId(productIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -108,7 +108,7 @@ final class InAppConfigStub {
     func getProductID_endsWith() -> InAppConfigResponse {
         let productIdTargeting = ProductIDTargeting(kind: .endsWith, value: "ots")
         let targeting = Targeting.viewProductId(productIdTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: nil,
+        let settings = Settings(operations: .init(viewProduct: nil,
                                                                       viewCategory: .init(systemName: "Hello"),
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -117,7 +117,7 @@ final class InAppConfigStub {
     func getProductSegment_Any() -> InAppConfigResponse {
         let productSegmentTargeting = ProductSegmentTargeting(kind: .positive, segmentationInternalId: "1", segmentationExternalId: "2", segmentExternalId: "3")
         let targeting = Targeting.viewProductSegment(productSegmentTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: .init(systemName: "Hello"),
+        let settings = Settings(operations: .init(viewProduct: .init(systemName: "Hello"),
                                                                       viewCategory: nil,
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
@@ -126,7 +126,7 @@ final class InAppConfigStub {
     func getProductSegment_None() -> InAppConfigResponse {
         let productSegmentTargeting = ProductSegmentTargeting(kind: .negative, segmentationInternalId: "1", segmentationExternalId: "2", segmentExternalId: "3")
         let targeting = Targeting.viewProductSegment(productSegmentTargeting)
-        let settings = InAppConfigResponse.Settings(operations: .init(viewProduct: .init(systemName: "Hello"),
+        let settings = Settings(operations: .init(viewProduct: .init(systemName: "Hello"),
                                                                       viewCategory: nil,
                                                                       setCart: nil))
         return .init(inapps: getInapps(with: targeting), monitoring: nil, settings: settings)
